requirements.txt
streamlit==1.21.0
matlab.engine == 23.2.1
pandas==1.5.3
numpy==1.23.5
Pillow==9.4.0
